requirements.txt
openai~=1.51.2
faiss-cpu
python-dotenv~=1.0.1
pydantic
pandas
fastapi~=0.115.2
pdfplumber~=0.11.4
numpy<2
transformers
torch
spacy
thinc